feat(admin): moved js from the view to the checkout.js file
